requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
Pillow==10.1.0
qrcode==7.4.2
python-engineio==4.7.1
Werkzeug==2.3.7